+++ src/components/Banner/Banner.jsx
@@ -1,13 +1,12 @@
+import $Banner from './Banner.styles.jsx';
 import Logo from './Logo/Logo.jsx';
 import Navigation from './Navigation/Navigation.jsx';
-import logo from '../../assets/carbon-tracker-logo.svg';
-import './Banner.css';
 
 export default function Banner () {
     return (
-        <header>
+        <$Banner>
             <Logo />
             <Navigation />
-        </header>
+        </$Banner>
     );
 }