+++ src/components/Banner/Banner.styles.jsx
@@ -1,0 +1,11 @@
+import { styled } from "styled-components";
+
+export const $Banner = styled.header`
+  position: absolute;
+  width: 100%;
+  height: 100px;
+  background-color: #ffffff;
+  box-shadow: 0 0 10px 0px rgba(0, 0, 0, 0.5);
+`;
+
+export default $Banner;