+++ src/components/Banner/Logo/Logo.jsx
@@ -1,8 +1,8 @@
+import $HeaderLogo from './Logo.styles.jsx'
 import logo from '../../../assets/carbon-tracker-logo.svg';
-import './Logo.css';
 
 export default function Logo() {
     return (
-        <img src={logo} alt="Carbon Tracker"/>
+        <$HeaderLogo src={logo} alt="Carbon Tracker"/>
     );
 }