+++ src/components/Banner/Logo/Logo.styles.jsx
@@ -1,0 +1,8 @@
+import { styled } from "styled-components";
+
+export const $HeaderLogo = styled.img`
+  height: 90%;
+  margin: 5px 100px;
+`;
+
+export default $HeaderLogo;