+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+0.12 (2024-05-04)
+===============
+* [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/2)
+* Implement styled-components instead .css files
+* Add tooltips for navigation menu items
+* Create and implement Navigation context
+* Add selected menu info to navigation
+
 0.11 (2024-05-02)
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/1)